fix: domain 모듈 분리 각 모듈로 나눔
