+++ libs/auth/src/index.ts
@@ -1,0 +1,2 @@
+export * from './auth.module';
+export * from './auth.service';
