+++ libs/auth/src/jwt.strategy.ts
