--- libs/domain/src/domain.module.ts
@@ -1,8 +1,0 @@
-import { Module } from '@nestjs/common';
-import { DomainService } from './domain.service';
-
-@Module({
-  providers: [DomainService],
-  exports: [DomainService],
-})
-export class DomainModule {}
