--- libs/domain/src/index.ts
@@ -1,2 +1,0 @@
-export * from './domain.module';
-export * from './domain.service';
