+++ libs/product/src/domain/product.ts
