+++ libs/product/src/dto/create-product.dto.ts
