+++ libs/product/src/index.ts
@@ -1,0 +1,2 @@
+export * from './product.module';
+export * from './product.service';
