+++ libs/product/src/product.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
-import { ProductController } from './product.controller';
+import { ProductController } from '@api/contorller/product.controller';
 import { ProductService } from './product.service';
 import { StorageModule } from '@app/storage';
-import { AuthModule } from '@api/auth/auth.module';
+import { AuthModule } from '@app/auth/auth.module';
 
 @Module({
   imports: [StorageModule, AuthModule],
