+++ libs/storage/src/mappers/user.mapper.ts
@@ -1,5 +1,5 @@
 import { UserEntity } from '@app/storage/entities/user.entity';
-import { User } from '@app/domain/user/user';
+import { User } from '@app/user/domain/user';
 
 export class UserMapper {
   static toDomain(entity: UserEntity): User {
