+++ libs/storage/src/storage.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { USER_REPOSITORY } from '@app/domain/user/user.repository';
-import { PRODUCT_REPOSITORY } from '@app/domain/catalog/product.repository';
+import { USER_REPOSITORY } from '@app/user/domain/user.repository';
+import { PRODUCT_REPOSITORY } from '@app/product/domain/product.repository';
 import { UserEntity } from '@app/storage/entities/user.entity';
 import { UserOrmRepository } from '@app/storage/repositories/user.orm.repository';
 import { ProductOrmRepository } from '@app/storage/repositories/product.orm.repository';
