+++ libs/user/src/index.ts
@@ -1,0 +1,2 @@
+export * from './user.module';
+export * from './user.service';
