+++ nest-cli.json
@@ -44,5 +44,34 @@
     "compilerOptions": {
       "tsConfigPath": "libs/core-enum/tsconfig.lib.json"
     }
+  },
+  "projects": {
+    "user": {
+      "type": "library",
+      "root": "libs/user",
+      "entryFile": "index",
+      "sourceRoot": "libs/user/src",
+      "compilerOptions": {
+        "tsConfigPath": "libs/user/tsconfig.lib.json"
+      }
+    },
+    "catalog": {
+      "type": "library",
+      "root": "libs/catalog",
+      "entryFile": "index",
+      "sourceRoot": "libs/catalog/src",
+      "compilerOptions": {
+        "tsConfigPath": "libs/catalog/tsconfig.lib.json"
+      }
+    },
+    "auth": {
+      "type": "library",
+      "root": "libs/auth",
+      "entryFile": "index",
+      "sourceRoot": "libs/auth/src",
+      "compilerOptions": {
+        "tsConfigPath": "libs/auth/tsconfig.lib.json"
+      }
+    }
   }
 }
